fix: resolve all ruff linting errors and frontend test failures
Backend:
- Fix undefined names in financials.py (AccountUpdate, DUTCH_RGS_LIGHT)
- Add _project_for_user_or_404 helper for cost endpoints
- Add missing import for customer email notifications
- Consolidate mid-file imports to top of financials.py
- Fix ambiguous Unicode chars (en-dash, multiplication sign)
- Fix line length violations
- Update ruff config: relax TCH rules for runtime-dependent modules,
  ignore false-positive style rules that conflict with FastAPI patterns
- Auto-fix import sorting, unused imports, enum modernization

Frontend:
- Fix push-permission.test.tsx: rewrite to match actual component API
- Fix push-notifications.test.tsx: keep PushManager alive through async renders
- Fix auth-refresh.test.tsx: add missing component mocks (offline-indicator,
  pwa-register, mobile-nav, mobile-time-tracker)
- Fix dashboard.test.tsx: mock apiFetch for Settings page to prevent
  unhandled ECONNREFUSED rejections

Results: ruff 0 errors, vitest 781/781 pass, pytest 729 pass (13 pre-existing)

+++ src/backend/app/services/planning/historical_learning.py
@@ -167,7 +167,7 @@ class HistoricalLearner:
 
         ratios = [r.actual_duration_s / r.estimated_duration_s for r in rows]
         bias = statistics.median(ratios)
-        corrected = int(round(estimated_duration_s * bias))
+        corrected = round(estimated_duration_s * bias)
         return Prediction(
             duration_s=corrected,
             confidence=_confidence(len(rows)),
